requirements.txt
Click>=6.0
dask>=1.2.2
dipy>=1.0.0
nipype>=1.2.0
pandas>=0.24.2
parse>=1.12.0
tqdm>=4.32.1
pybids>=0.8.0
matplotlib>=3.1.0
cytoolz>=0.9.0.1
numba>=0.43.1
